Added weapon switching, gun graphics
Weapon switching can be used by pressing 1 (for primary weapon) and 2 (for secondary weapon). Note that the current primary and secondary models are the same, so only the damage done will appear to be different. Gun graphics were implemented using Unity's particle system.

BUGS TO FIX: Leaning does not reset on respawn; CharacterController collider currently acts as a hitbox (that apparently cannot be disabled) for raycasts, but we would rather have hitboxes be based on the player model.