Added a scoped-in overlay on zoom, created a less-tedious interface to adding new float-based settings
The scoped-in overlay is a UI image that is enabled on zoom and disabled on unzoom. The player's viewmodel is also disabled/enabled accordingly. To add a float-based setting, we must add a new group of setting members in the PlayerSettings script, add a FloatEditor prefab into the scene, and then reference the prefab in the PlayerSettings script.